Initial steps of member table filtering
