+++ src/main/webapp/entries.js
@@ -81,7 +81,6 @@ module.exports = {
   "sequencing-runs":
     "./resources/js/pages/sequencing-runs/SequencingRunsPage.jsx",
   groups: "./resources/js/pages/UserGroupsPage",
-  "group-details": "./resources/js/pages/UserGroupDetailsPage",
   "ncbi-exports": "./resources/js/pages/projects/ncbi-export.js",
   search: "./resources/js/pages/search/search.js",
   "run-files": "./resources/js/pages/sequence-files/run-files.js",
