--- src/main/webapp/pages/groups/details.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
-	  data-layout-decorate="~{template/page}">
-<head>
-<title th:text="${group.label}">_IRIDA Group_</title>
-	<link rel="stylesheet" th:href="@{/dist/css/group-details.bundle.css}">
-</head>
-<body>
-<div layout:fragment="page" id="group-root" style="height: 100%">
-</div>
-
-<th:block layout:fragment="scripts">
-	<script th:src="@{/dist/js/group-details.bundle.js}"></script>
-</th:block>
-</body>
-</html>
