+++ src/main/webapp/resources/js/utilities/table-utilities.js
@@ -1,0 +1,4 @@
+export const stringSorter = (property) => (a, b) =>
+  a[property].localeCompare(b[property], window.TL.LANGUAGE_TAG, {
+    sensitivity: "base",
+  });
